refactor(terrain): extract shared utility functions to terrainUtils.ts
